Changing flood control to be seconds and allowing it to same User
Updating initial topic creation permissions
